[SWITCHYARD-120] - Moved test-util to test and to package org.switchyard.test
