+++ test/src/main/java/org/switchyard/test/MockHandler.java
@@ -19,9 +19,13 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.switchyard;
+package org.switchyard.test;
 
 import java.util.concurrent.LinkedBlockingQueue;
+
+import org.switchyard.BaseHandler;
+import org.switchyard.Exchange;
+import org.switchyard.HandlerException;
 
 import junit.framework.TestCase;
