+++ test/src/main/java/org/switchyard/test/MockInitialContextFactory.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.switchyard;
+package org.switchyard.test;
 
 import javax.naming.Context;
 import javax.naming.NamingException;
