--- test/test-util/pom.xml
@@ -1,89 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ JBoss, Home of Professional Open Source.
-  ~ Copyright 2010, Red Hat, Inc., and individual contributors
-  ~ as indicated by the @author tags. See the copyright.txt file in the
-  ~ distribution for a full listing of individual contributors.
-  ~
-  ~ This is free software; you can redistribute it and/or modify it
-  ~ under the terms of the GNU Lesser General Public License as
-  ~ published by the Free Software Foundation; either version 2.1 of
-  ~ the License, or (at your option) any later version.
-  ~
-  ~ This software is distributed in the hope that it will be useful,
-  ~ but WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-  ~ Lesser General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU Lesser General Public
-  ~ License along with this software; if not, write to the Free
-  ~ Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
-  ~ 02110-1301 USA, or see the FSF site: http://www.fsf.org.
-  -->
-
-<project
-	xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    
-	<parent>
-		<groupId>org.switchyard</groupId>
-		<artifactId>switchyard-core-parent</artifactId>
-		<version>1.0-SNAPSHOT</version>
-		<relativePath>../../pom.xml</relativePath>
-	</parent>
-	
-	<groupId>org.switchyard</groupId>
-	<artifactId>switchyard-test-util</artifactId>
-	
-	<packaging>jar</packaging>
-	<name>Switchyard: Test Utilities</name>
-	<version>${project.parent.version}</version>
-	<description>Test Utility classes.</description>
-	<url>http://switchyard.org/</url>
-
-    <properties>
-        <weld.version>1.1.0-SNAPSHOT</weld.version>
-    </properties>
-
-	<dependencies>
-		<!-- internal dependencies -->
-        <dependency>
-            <groupId>org.switchyard</groupId>
-            <artifactId>switchyard-runtime</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.switchyard</groupId>
-            <artifactId>switchyard-deploy</artifactId>
-        </dependency>
-		<!-- N/A -->
-		<!-- external dependencies -->
-        <dependency>
-            <groupId>log4j</groupId>
-            <artifactId>log4j</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>xmlunit</groupId>
-            <artifactId>xmlunit</artifactId>
-        </dependency>
-
-        <dependency>
-           <groupId>org.jboss.weld.se</groupId>
-           <artifactId>weld-se-core</artifactId>
-            <version>${weld.version}</version>
-        </dependency>
-
-        <dependency>
-        	<groupId>org.slf4j</groupId>
-        	<artifactId>slf4j-simple</artifactId>
-        	<version>1.6.1</version>
-        	<scope>runtime</scope>
-        </dependency>
-
-    </dependencies>
-</project>
