+++ tools/maven/archetypes/application/src/main/resources/archetype-resources/pom.xml
@@ -21,7 +21,7 @@
     </dependency>
     <dependency>
       <groupId>org.switchyard</groupId>
-      <artifactId>switchyard-test-util</artifactId>
+      <artifactId>switchyard-test</artifactId>
       <version>1.0-SNAPSHOT</version>
     </dependency>
     <dependency>
